--- run_longsd.sh
@@ -1,2 +1,0 @@
-CUDA_VISIBLE_DEVICES=1 python longsd.py -f "$1" -n 2 &
-CUDA_VISIBLE_DEVICES=0 python longsd.py -f "$1" -n 2 && fg
